Refactor and polish codes.
