+++ components/raftstore/src/store/fsm/store.rs
@@ -103,8 +103,8 @@ use crate::{
         util,
         util::{is_initial_msg, RegionReadProgressRegistry},
         worker::{
-            Assistant, AutoSplitController, CleanupRunner, CleanupSstRunner, CleanupSstTask,
-            CleanupTask, CompactRunner, CompactTask, ConsistencyCheckRunner, ConsistencyCheckTask,
+            AutoSplitController, CleanupRunner, CleanupSstRunner, CleanupSstTask, CleanupTask,
+            CompactRunner, CompactTask, ConsistencyCheckRunner, ConsistencyCheckTask,
             GcSnapshotRunner, GcSnapshotTask, PdRunner, RaftlogGcRunner, RaftlogGcTask,
             ReadDelegate, RefreshConfigRunner, RefreshConfigTask, RegionRunner, RegionTask,
             SplitCheckTask,
@@ -1597,8 +1597,6 @@ struct Workers<EK: KvEngine, ER: RaftEngine> {
     coprocessor_host: CoprocessorHost<EK>,
 
     refresh_config_worker: LazyWorker<RefreshConfigTask>,
-
-    assistant: Assistant<EK, ER>,
 }
 
 pub struct RaftBatchSystem<EK: KvEngine, ER: RaftEngine> {
@@ -1691,11 +1689,6 @@ impl<EK: KvEngine, ER: RaftEngine> RaftBatchSystem<EK, ER> {
             raftlog_fetch_worker: Worker::new("raftlog-fetch-worker"),
             coprocessor_host: coprocessor_host.clone(),
             refresh_config_worker: LazyWorker::new("refreash-config-worker"),
-            assistant: {
-                let mut assistant = Assistant::new("engine-assistant", engines.clone());
-                assistant.start();
-                assistant
-            },
         };
         mgr.init()?;
         let region_runner = RegionRunner::new(
@@ -1920,9 +1913,6 @@ impl<EK: KvEngine, ER: RaftEngine> RaftBatchSystem<EK, ER> {
             return;
         }
         let mut workers = self.workers.take().unwrap();
-        // Stop all background workers in the engine before stopping
-        // the other workers.
-        workers.assistant.stop();
         // Wait all workers finish.
         workers.pd_worker.stop();
