+++ components/server/src/server.rs
@@ -1567,8 +1567,18 @@ where
         }
     }
 
+    fn prepare_stop(&self) {
+        self.engines.as_ref().map(|engines| {
+            // Disable manul compaction jobs before shutting down the engines. And it
+            // will stop the compaction thread in advance, so it won't block the
+            // cleanup thread when exiting.
+            let _ = engines.engines.kv.disable_manual_compaction();
+        });
+    }
+
     fn stop(self) {
         tikv_util::thread_group::mark_shutdown();
+        self.prepare_stop();
         let mut servers = self.servers.unwrap();
         servers
             .server
